+++ hyperspy/misc/tv_denoise.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright 2007-2020 The HyperSpy developers
+# Copyright 2007-2021 The HyperSpy developers
 #
 # This file is part of  HyperSpy.
 #
@@ -43,8 +43,7 @@ def _tv_denoise_3d(im, weight=100, eps=2.e-4, keep_type=False, n_iter_max=200):
 
     keep_type: bool, optional (False)
         whether the output has the same dtype as the input array.
-        keep_type is False by default, and the dtype of the output
-        is np.float
+        keep_type is False by default, and the dtype of the output is float
 
     n_iter_max: int, optional
         maximal number of iterations used for the optimization.
@@ -64,13 +63,13 @@ def _tv_denoise_3d(im, weight=100, eps=2.e-4, keep_type=False, n_iter_max=200):
 
     >>> x, y, z = np.ogrid[0:40, 0:40, 0:40]
     >>> mask = (x -22)**2 + (y - 20)**2 + (z - 17)**2 < 8**2
-    >>> mask = mask.astype(np.float)
+    >>> mask = mask.astype(float)
     >>> mask += 0.2*np.random.randn(*mask.shape)
     >>> res = tv_denoise_3d(mask, weight=100)
     """
     im_type = im.dtype
-    if im_type is not np.float:
-        im = im.astype(np.float)
+    if im_type is not float:
+        im = im.astype(float)
     px = np.zeros_like(im)
     py = np.zeros_like(im)
     pz = np.zeros_like(im)
@@ -137,8 +136,7 @@ def _tv_denoise_2d(im, weight=50, eps=2.e-4, keep_type=False, n_iter_max=200):
 
     keep_type: bool, optional (False)
         whether the output has the same dtype as the input array.
-        keep_type is False by default, and the dtype of the output
-        is np.float
+        keep_type is False by default, and the dtype of the output is float
 
     n_iter_max: int, optional
         maximal number of iterations used for the optimization.
@@ -166,13 +164,13 @@ def _tv_denoise_2d(im, weight=50, eps=2.e-4, keep_type=False, n_iter_max=200):
     Examples
     ---------
     >>> import scipy
-    >>> ascent = scipy.ascent().astype(np.float)
+    >>> ascent = scipy.ascent().astype(float)
     >>> ascent += 0.5 * ascent.std()*np.random.randn(*ascent.shape)
     >>> denoised_ascent = tv_denoise(ascent, weight=60.0)
     """
     im_type = im.dtype
-    if im_type is not np.float:
-        im = im.astype(np.float)
+    if im_type is not float:
+        im = im.astype(float)
     px = np.zeros_like(im)
     py = np.zeros_like(im)
     gx = np.zeros_like(im)
@@ -233,8 +231,7 @@ def _tv_denoise_1d(im, weight=50, eps=2.e-4, keep_type=False, n_iter_max=200):
 
     keep_type: bool, optional (False)
         whether the output has the same dtype as the input array.
-        keep_type is False by default, and the dtype of the output
-        is np.float
+        keep_type is False by default, and the dtype of the output is float
 
     n_iter_max: int, optional
         maximal number of iterations used for the optimization.
@@ -262,13 +259,13 @@ def _tv_denoise_1d(im, weight=50, eps=2.e-4, keep_type=False, n_iter_max=200):
     Examples
     ---------
     >>> import scipy
-    >>> ascent = scipy.misc.ascent().astype(np.float)
+    >>> ascent = scipy.misc.ascent().astype(float)
     >>> ascent += 0.5 * ascent.std()*np.random.randn(*ascent.shape)
     >>> denoised_ascent = tv_denoise(ascent, weight=60.0)
     """
     im_type = im.dtype
-    if im_type is not np.float:
-        im = im.astype(np.float)
+    if im_type is not float:
+        im = im.astype(float)
     px = np.zeros_like(im)
     gx = np.zeros_like(im)
     d = np.zeros_like(im)
@@ -325,8 +322,7 @@ def tv_denoise(im, weight=50, eps=2.e-4, keep_type=False, n_iter_max=200):
 
     keep_type: bool, optional (False)
         whether the output has the same dtype as the input array.
-        keep_type is False by default, and the dtype of the output
-        is np.float
+        keep_type is False by default, and the dtype of the output is float
 
     n_iter_max: int, optional
         maximal number of iterations used for the optimization.
@@ -361,13 +357,13 @@ def tv_denoise(im, weight=50, eps=2.e-4, keep_type=False, n_iter_max=200):
     ---------
     >>> # 2D example using ascent
     >>> import scipy
-    >>> ascent = scipy.misc.ascent().astype(np.float)
+    >>> ascent = scipy.misc.ascent().astype(float)
     >>> ascent += 0.5 * ascent.std()*np.random.randn(*ascent.shape)
     >>> denoised_ascent = tv_denoise(ascent, weight=60)
     >>> # 3D example on synthetic data
     >>> x, y, z = np.ogrid[0:40, 0:40, 0:40]
     >>> mask = (x -22)**2 + (y - 20)**2 + (z - 17)**2 < 8**2
-    >>> mask = mask.astype(np.float)
+    >>> mask = mask.astype(float)
     >>> mask += 0.2*np.random.randn(*mask.shape)
     >>> res = tv_denoise_3d(mask, weight=100)
     """
